[이태환] 데미지버그수정

+++ Assets/02.Scripts/Enemy/Monster_Bat.cs
@@ -269,7 +269,7 @@ public class Monster_Bat : MonoBehaviourPun, IPunObservable, IDamaged
     }
 
     [PunRPC]
-    public void Damaged(int damage, int actorNumber)
+    public void Damaged(float damage, int actorNumber)
     {
         if (state == MonsterState.Death || !PhotonNetwork.IsMasterClient)
         {
